refactor: polish request-transformer plugin to match project conventions
Great foundation from the initial implementation! This commit brings
the plugin in line with existing conventions and addresses a few
edge cases:

- Consolidate source files into single lib.rs (consistent with all
  other plugins)
- Snapshot original request before transforms so that variable
  interpolation ($query.*, $header.*, etc.) always resolves against
  the incoming request — this makes the query-to-body use case from
  ADR-0020 work correctly
- Cache compiled regex in a #[serde(skip)] field to avoid
  recompilation on every request
- Preserve JSON types for interpolated values in body transforms
  (e.g. $query.page=2 becomes the number 2, not the string "2")
- Use #[cfg(target_arch = "wasm32")] for host function bindings
  with mock_host module for native tests
- Add "rlib" to crate-type and "additionalProperties": false to
  config-schema.json

51 tests passing (including new snapshot isolation, lazy regex, and
context interpolation tests).
