+++ plugins/request-transformer/Cargo.toml
@@ -9,7 +9,7 @@ license = "MIT"
 [workspace]
 
 [lib]
-crate-type = ["cdylib"]
+crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 barbacane-plugin-sdk = { path = "../../crates/barbacane-plugin-sdk" }
