+++ plugins/request-transformer/config-schema.json
@@ -101,5 +101,6 @@
         }
       }
     }
-  }
+  },
+  "additionalProperties": false
 }
